requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Hugging Face
datasets>=2.14.0
transformers>=4.30.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Metrics
scikit-learn>=1.3.0

# Progress bars
tqdm>=4.65.0

# For GCP compatibility
google-cloud-storage>=2.10.0
